Removed PipingProbabilityAssessmentInput property from PipingCalculation Related to Issue [WTI-468]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@3470 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: b4a8aa29c94feedc0c32c6187aea543d07324dbb
Former-commit-id: 3ea92116d9b14e96908618d5332108d86422d91c

+++ Ringtoets/Common/src/Ringtoets.Common.Service/ProbabilityAssessmentService.cs
@@ -37,7 +37,7 @@ namespace Ringtoets.Common.Service
         /// to the total of the failure probability of the assessment section.</param>
         /// <param name="lengthEffectN">The 'N' parameter used to factor in the 'length effect'.</param>
         /// <param name="reliability">The reliability to use for the calculation.</param>
-        public static ProbabilityAssessmentOutput Calculate(double norm, double contribution, double lengthEffectN, double reliability)
+        public static ProbabilityAssessmentOutput Calculate(int norm, double contribution, double lengthEffectN, double reliability)
         {
             var requiredProbability = RequiredProbability(contribution / 100.0, norm, lengthEffectN);
             var probability = ReliabilityToProbability(reliability);
@@ -51,9 +51,9 @@ namespace Ringtoets.Common.Service
                                                    factorOfSafety);
         }
 
-        private static double RequiredProbability(double contribution, double norm, double lengthEffectN)
+        private static double RequiredProbability(double contribution, int norm, double lengthEffectN)
         {
-            return contribution*(1/norm)/lengthEffectN;
+            return contribution*(1.0/norm)/lengthEffectN;
         }
 
         private static double ReliabilityToProbability(double reliability)
